+++ src/deltachat_cursed/event.py
@@ -15,24 +15,17 @@ class AccountPlugin:
     def __init__(self, account: Account) -> None:
         self.account = account
         self.chatlist_monitors: Set[ChatListMonitor] = set()
-        self.current_chat = self.get_chats()[0]
-
-    def get_chats(self) -> list:
-        chats = self.account.get_chats()
-        deaddrop = self.account.get_deaddrop_chat()
-        if deaddrop not in chats:
-            chats.insert(0, deaddrop)
-        return chats
+        self.current_chat = self.account.get_chats()[0]
 
     def add_chatlist_monitor(self, monitor: ChatListMonitor) -> None:
         self.chatlist_monitors.add(monitor)
-        monitor.chatlist_changed(self.get_current_index(), self.get_chats())
+        monitor.chatlist_changed(self.get_current_index(), self.account.get_chats())
 
     def remove_monitor(self, monitor: ChatListMonitor) -> None:
         self.chatlist_monitors.discard(monitor)
 
     def chatlist_changed(self) -> None:
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         if self.current_chat not in chats:
             if chats:
                 self.current_chat = chats[0]
@@ -47,13 +40,13 @@ class AccountPlugin:
             m.chatlist_changed(index, chats)
 
     def select_chat(self, index: Optional[int]) -> None:
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         for m in self.chatlist_monitors:
             m.chat_selected(index, chats)
         self.current_chat = None if index is None else chats[index]
 
     def select_chat_by_id(self, chat_id: int) -> None:
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         chat = self.account.get_chat_by_id(chat_id)
         index = chats.index(chat)
         for m in self.chatlist_monitors:
@@ -61,7 +54,7 @@ class AccountPlugin:
         self.current_chat = chat
 
     def select_next_chat(self) -> None:
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         if self.current_chat is None:
             if chats:
                 self.select_chat(len(chats) - 1)
@@ -73,7 +66,7 @@ class AccountPlugin:
             self.select_chat(i)
 
     def select_previous_chat(self) -> None:
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         if self.current_chat is None:
             if chats:
                 self.select_chat(0)
@@ -87,7 +80,7 @@ class AccountPlugin:
     def get_current_index(self) -> Optional[int]:
         if self.current_chat is None:
             return None
-        chats = self.get_chats()
+        chats = self.account.get_chats()
         return chats.index(self.current_chat)
 
     @account_hookimpl